requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
strawberry-graphql[fastapi]>=0.220.0
sqlalchemy>=2.0.0
